// infinite craft/src/utils/enums/colors.ts
export enum Colors {
  WHITE = "#FFFFFF",
  LAVENDER = "#E5EFFE",
  ZOMP = "#2D969B",
  ANTI_FLASH_WHITE = "#F2F2F2",
  GAINSBORO = "#DCDCDC",
  CHINESE_SILVER = "#c9c7c7",
  GHOST_WHITE = "#F8FBFF",
  SONIC_SILVER = "#7A7A7A",
  LIGHT_STEEL_BLUE = "#ADBFDA",
  CULTURED = "#F4F4F4",
  BLACK = "#000000",
  ALICE_BLUE = "#EEF5FF",
  RAISIN_BLACK = "#262626",
  BLACK_CORAL = "#5A5E6B",
  PHILIPPINE_SILVER = "#B6B3B3",
  SPANISH_GRAY = "#9A9A9A",
  BLUE_JEANS = "#60a5fa",
  RGBA_25 = "#2D969B",
  DARK_JUNGLE_GREEN = "#161F26",
  BLACK_OLIVE = "#423F3F",
}